add a new middleware that following the latest Vercel APIs
